+++ CHANGES.md
@@ -56,3 +56,4 @@
 * support for tls 0.5.0
 * add hybrid irmin/workdir storage - messages are stored in workdir on FS, metadata is stored in Git
 * add multiple domain support (virtual user account folder for user@domain is domain/user)
+* add COMPRESS=DEFLATE capability (rfc 4978)